+++ src/Symfony/Component/Messenger/Transport/Serialization/TypeResolver/TypeResolverInterface.php
@@ -1,0 +1,22 @@
+<?php
+
+/*
+ * This file is part of the Symfony package.
+ *
+ * (c) Fabien Potencier <fabien@symfony.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Symfony\Component\Messenger\Transport\Serialization\TypeResolver;
+
+/**
+ * Resolve denormalization class based on message headers.
+ *
+ * @author Laurent VOULLEMIER <laurent.voullemier@gmail.com>
+ */
+interface TypeResolverInterface
+{
+    public function resolve(array $encodedEnvelope): string;
+}